+++ semana19/projeto/template-labook/.gitignore
@@ -1,0 +1,5 @@
+node_modules
+build
+.vscode
+.env
+.rest
