+++ semana19/projeto/template-labook/package.json
@@ -5,8 +5,9 @@
    "main": "index.js",
    "scripts": {
       "migrations": "tsc && node ./build/migrations.js",
-      "start": "npm run build && node --inspect ./build/src/index.js",
-      "dev": "clear && ts-node-dev ./src/index.ts",
+      
+      "start": "tsc && node --inspect ./build/index.js",
+      "dev": "ts-node-dev ./src/index.ts",
       "build": "clear && echo \"Transpiling files...\" && tsc && echo \"Done!\" ",
       "test": "jest"
    },
