+++ semana19/projeto/template-labook/src/controller/app.ts
@@ -1,0 +1,13 @@
+import express from 'express'
+import cors from 'cors'
+
+const app = express()
+
+app.use(express.json())
+app.use(cors())
+
+app.listen(3003, () => {
+   console.log('Server running on port 3003.')
+})
+
+export default app
