+++ semana19/projeto/template-labook/src/data/BaseDatabase.ts
@@ -1,0 +1,20 @@
+import knex from 'knex'
+import dotenv from 'dotenv'
+
+dotenv.config();
+
+export default class BaseDatabase{
+
+   protected static connection = knex({
+      client: 'mysql',
+      connection: {
+         host: process.env.DB_HOST,
+         user: process.env.DB_USER,
+         password: process.env.DB_PASSWORD,
+         database: process.env.DB_SCHEMA,
+         port: 3306,
+         multipleStatements: true
+      }
+   })
+}
+
