+++ semana19/projeto/template-labook/src/index.ts
@@ -8,32 +8,37 @@ import * as jwt from "jsonwebtoken"
 import * as bcrypt from "bcryptjs"
 import { v4 } from "uuid"
 import Knex from "knex"
+import BaseDatabase from "./data/BaseDatabase"
 
 /**************************** CONFIG ******************************/
 
-dotenv.config()
+// dotenv.config()
 
-export const connection: Knex = knex({
-   client: "mysql",
-   connection: {
-      host: process.env.DB_HOST,
-      user: process.env.DB_USER,
-      password: process.env.DB_PASSWORD,
-      database: process.env.DB_SCHEMA,
-      port: 3306,
-      multipleStatements: true
-   }
-})
+// export const connection: Knex = knex({
+//    client: "mysql",
+//    connection: {
+//       host: process.env.DB_HOST,
+//       user: process.env.DB_USER,
+//       password: process.env.DB_PASSWORD,
+//       database: process.env.DB_SCHEMA,
+//       port: 3306,
+//       multipleStatements: true
+//    }
+// })
+// data - BaseDatabase.ts
 
-const app: Express = express()
-app.use(express.json())
-app.use(cors())
+// const app: Express = express()
+// app.use(express.json())
+// app.use(cors())
+// controller - app.ts
 
 /**************************** TYPES ******************************/
 
-type authenticationData = {
-   id: string
-}
+// type authenticationData = {
+//    id: string
+// }
+// model - authenticationData.ts
+
 
 type user = {
    id: string,
@@ -58,44 +63,47 @@ type post = {
 
 /**************************** SERVICES ******************************/
 
-const generateId = (): string => v4()
+// const generateId = (): string => v4()
+// services - idGenerator
 
-function generateToken(
-   payload: authenticationData
-): string {
-   return jwt.sign(
-      payload,
-      process.env.JWT_KEY as string,
-      {
-         expiresIn: process.env.JWT_EXPIRES_IN
-      }
-   )
-}
+// function generateToken(
+//    payload: authenticationData
+// ): string {
+//    return jwt.sign(
+//       payload,
+//       process.env.JWT_KEY as string,
+//       {
+//          expiresIn: process.env.JWT_EXPIRES_IN
+//       }
+//    )
+// }
 
-function getTokenData(
-   token: string
-): authenticationData {
-   const result: any = jwt.verify(
-      token,
-      process.env.JWT_KEY as string
-   )
+// function getTokenData(
+//    token: string
+// ): authenticationData {
+//    const result: any = jwt.verify(
+//       token,
+//       process.env.JWT_KEY as string
+//    )
 
-   return { id: result.id, }
-}
+//    return { id: result.id, }
+// }
+// services - authenticator.ts
 
-const hash = async (
-   plainText: string
-): Promise<string> => {
-   const rounds = Number(process.env.BCRYPT_COST);
-   const salt = await bcrypt.genSalt(rounds);
-   return bcrypt.hash(plainText, salt)
-}
+// const hash = async (
+//    plainText: string
+// ): Promise<string> => {
+//    const rounds = Number(process.env.BCRYPT_COST);
+//    const salt = await bcrypt.genSalt(rounds);
+//    return bcrypt.hash(plainText, salt)
+// }
 
-const compare = async (
-   plainText: string, cypherText: string
-): Promise<boolean> => {
-   return bcrypt.compare(plainText, cypherText)
-}
+// const compare = async (
+//    plainText: string, cypherText: string
+// ): Promise<boolean> => {
+//    return bcrypt.compare(plainText, cypherText)
+// }
+// services - hashManager.ts
 
 /**************************** ENDPOINTS ******************************/
 
@@ -253,6 +261,7 @@ app.get('/posts/:id', async (req: Request, res: Response) => {
 
 /**************************** SERVER INIT ******************************/
 
-app.listen(3003, () => {
-   console.log("Server running on port 3003")
-})
+// app.listen(3003, () => {
+//    console.log("Server running on port 3003")
+// }) 
+// controller - app.ts
