+++ semana19/projeto/template-labook/src/model/authenticationData.ts
@@ -1,0 +1,3 @@
+export type authenticationData = {
+    id: string
+ }
