+++ semana19/projeto/template-labook/src/services/authenticator.ts
@@ -1,0 +1,27 @@
+import { authenticationData } from "../model/authenticationData"
+import * as jwt from "jsonwebtoken"
+import dotenv from 'dotenv'
+
+export function generateToken(
+    payload: authenticationData
+ ): string {
+    return jwt.sign(
+       payload,
+       process.env.JWT_KEY as string,
+       {
+          expiresIn: process.env.JWT_EXPIRES_IN
+       }
+    )
+ }
+ 
+ export function getTokenData(
+    token: string
+ ): authenticationData {
+    const result: any = jwt.verify(
+       token,
+       process.env.JWT_KEY as string
+    )
+ 
+    return { id: result.id, }
+ }
+
