+++ semana19/projeto/template-labook/src/services/idGenerator.ts
@@ -1,0 +1,3 @@
+import { v4 } from "uuid"
+
+export const generateId = (): string => v4()
